+++ src/ecstacy/screens/modals/__init__.py
@@ -6,12 +6,14 @@ from ecstacy.screens.modals.chart_mapping import (
 from ecstacy.screens.modals.column_picker import ColumnPickerScreen
 from ecstacy.screens.modals.export import ExportScreen
 from ecstacy.screens.modals.open import OpenScreen
+from ecstacy.screens.modals.theme_picker import ThemePickerScreen
 
 __all__ = [
     "ChartMappingScreen",
     "ColumnPickerScreen",
     "ExportScreen",
     "OpenScreen",
+    "ThemePickerScreen",
     "VIZ_FIELDS",
     "VIZ_NO_MAPPING",
 ]